// frontend/src/app/core/models/theme.model.ts
export interface ThemeOption {
  key: string;
  name: string;

  colorPrincipal: string;
  colorSecundario: string;
  colorTercero: string;
  colorPaginaPrincipal: string;

  colorTextoPrimario: string;
  colorTextoSecundario: string;

  colorBorder: string;
  colorShadow: string;
  colorDanger: string;
  colorHover: string;

  colorGrafico1: string;
  colorGrafico2: string;
  colorGrafico3: string;
  colorGrafico4: string;
  colorGrafico5: string;
}